sparse: implement typeof_unqual()
The typeof_unqual() is a c23 operator which is similar to typeof() except
it masks away the qualifiers, ie. MOD_CONST, MOD_VOLATILE and
MOD_RESTRICT.  So the implementation of typeof() and typeof_unqual() is
basically the same except that in examine_typeof_helper() we have the mask:

        mod &= ~MOD_QUALIFIER;

We will still preserve the MOD_ATOMIC, MOD_NOCAST and MOD_SPECIFIER
modifiers with both typeof() and typeof_unqual().

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>
Signed-off-by: Chris Li <sparse@chrisli.org>